pyproject.toml
[tool.poetry]
name = "python_template"
version = "0.1.0"
description = ""
authors = ["ldom <louis@giron-dom.eu>"]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
mongodb-python-manager = "^1.7.1"
python-dotenv = "^1.0.0"
boto3 = "^1.34.18"
structlog = "23.3.0"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
pydantic = "^2.5.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
black = {extras = ["jupyter"], version = "^23.12.1"}
pre-commit = "^3.5.0"
isort = "^5.13.2"
pylint = "^3.0.3"
flake8 = "^7.0.0"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-design = "^0.5.0"
sphinx-rtd-theme = "^1.3.0"
autodocsumm = "^0.2.11"
ipython = "^8.17.2"
matplotlib = "^3.8.1"
pypandoc = "^1.12"
nbsphinx = "^0.9.3"
sphinx-bootstrap-theme = "^0.8.1"
sphinx-gallery = "^0.15.0"
jupytext = "^1.16.0"
jupyter = "^1.0.0"
json-schema-for-humans = "^0.47"
pickleshare = "^0.7.5"
pydata-sphinx-theme = "^0.15.1"
requests = "^2.31.0"
sphinxcontrib-openapi = "^0.8.3"
mistune = "^2.0.5"


[tool.poetry.group.tests.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
httpx = "^0.25.1"

[tool.poetry.group.deploy.dependencies]
wheel = "^0.41.3"
twine = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
